+++ src/Neo4jBoostServiceProvider.php
@@ -18,6 +18,8 @@ use Neo4j\LaravelBoost\Console\StartNeo4jCommand;
 use Neo4j\LaravelBoost\Console\TestStdioCommand;
 use Neo4j\LaravelBoost\ContainerGraph\AuthConfigExtractor;
 use Neo4j\LaravelBoost\ContainerGraph\BindingLifetimeResolver;
+use Neo4j\LaravelBoost\ContainerGraph\BroadcastChannelExtractor;
+use Neo4j\LaravelBoost\ContainerGraph\BroadcastConnectionExtractor;
 use Neo4j\LaravelBoost\ContainerGraph\ContextualBindingExtractor;
 use Neo4j\LaravelBoost\ContainerGraph\ContextualGiveResolver;
 use Neo4j\LaravelBoost\ContainerGraph\DependencyChainBuilder;
@@ -93,6 +95,8 @@ class Neo4jBoostServiceProvider extends ServiceProvider
         $this->app->singleton(QueueConnectionExtractor::class);
         $this->app->singleton(ScheduledTaskExtractor::class);
         $this->app->singleton(AuthConfigExtractor::class);
+        $this->app->singleton(BroadcastConnectionExtractor::class);
+        $this->app->singleton(BroadcastChannelExtractor::class);
         $this->app->singleton(LaravelFirstPartyFacadeCatalog::class);
         $this->app->singleton(FacadeCatalogExporter::class);
         $this->app->singleton(FacadeAccessorParser::class);